Setting Stable Base (#3)
* Update user_guide.md

Removed "Happy documenting!" due to it's unnecessary nature.

* Fixed Acknowledgements to readme.md

Changed the "[YOUR NAME]" to actually have acknowledgements.

* erd

* Fixed the introduction.

Removed the "software engineering requirements" phrase due to its repetitive nature and unnecessary description of "document".

* Added inspirations.

Added "java" and "YAML" entries to the inspirations section.

* Create philosophy.md

* docuflex classes

* Added long term vision document.

This document details our long-term goals and visions for what DocuFlex will become in the future as we build it. Adding it for more complete documentation.

---------

Co-authored-by: RawBoeuf <61438548+RawBoeuf@users.noreply.github.com>
Co-authored-by: = <=>